backend/pyproject.toml
[project]
name = "skillval-backend"
version = "0.1.0"
description = "SkillVal FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "motor>=3.4.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyjwt[crypto]>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=23.2.1",
    "redis>=5.0.5",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.4.5",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
